+++ trilby-cli/src/Trilby/Install/Config/Host.hs
@@ -2,25 +2,10 @@ module Trilby.Install.Config.Host where
 
 import Trilby.HNix
 import Trilby.Install.Config.Edition
+import Trilby.Install.Config.Keyboard
 import Trilby.Install.Config.Release
 import Trilby.Install.Config.User
 import Prelude
-
-data Keyboard = Keyboard
-    { layout :: Text
-    , variant :: Maybe Text
-    }
-    deriving stock (Generic, Show)
-
-instance ToExpr Keyboard where
-    toExpr Keyboard{..} =
-        [nix|
-        {
-            layout = layout;
-            variant = variant;
-        }
-        |]
-            & canonicalSet
 
 data Host = Host
     { hostname :: Text